refactor(types): introduce new types and clean up code structure
- Added a new `redmine.types.ts` file to define `ProjectUnit` and `FinancialData` interfaces, improving type organization and clarity.
- Updated `redmine.ts` to import types from `redmine.types.ts`, enhancing modularity and reducing redundancy.
- Removed commented-out threshold settings in `vitest.config.ts` for cleaner configuration.
- Refactored `weeklyFinancialReports` workflow to use an arrow function for better consistency in function definitions.

These changes enhance code maintainability and improve type safety across the project.
