+++ workers/main/src/activities/weeklyFinancialReports/index.ts
@@ -1,1 +1,2 @@
 export * from './redmine';
+export * from './redmine.types';
