+++ workers/main/vitest.config.ts
@@ -14,12 +14,6 @@ export default defineConfig({
       all: true,
       include: ['src/**/*.ts'],
       exclude: ['src/__tests__/**', 'src/dist/**'],
-      // thresholds: {
-      //   statements: 70,
-      //   branches: 70,
-      //   functions: 70,
-      //   lines: 70,
-      // },
     },
   },
 });
